multi kernel develop

+++ lightseq/csrc/layers_new/dec_enc_attention_layer.cpp
@@ -126,9 +126,9 @@ template <typename T1, typename T2>
 void DecEncAttentionLayer<T1, T2>::before_backward() {}
 
 template <typename T1, typename T2>
-int DecEncAttentionLayer<T1, T2>::load_para_and_grad(
+size_t DecEncAttentionLayer<T1, T2>::load_para_and_grad(
     const T1* para_ptr, T2* grad_ptr) {  // for training
-  int offset = 0;
+  size_t offset = 0;
   _attn_qw->set_value((char*)(para_ptr + offset));
   _attn_qw->set_grad((char*)(grad_ptr + offset));
   _attn_qw->set_shape({_hidden_size, _hidden_size});
