+++ lightseq/csrc/layers_new/gpt_attention_layer.cpp
@@ -135,9 +135,9 @@ template <typename T1, typename T2>
 void GptAttentionLayer<T1, T2>::before_backward() {}
 
 template <typename T1, typename T2>
-int GptAttentionLayer<T1, T2>::load_para_and_grad(
+size_t GptAttentionLayer<T1, T2>::load_para_and_grad(
     const T1* para_ptr, T2* grad_ptr) {  // for training
-  int offset = 0;
+  size_t offset = 0;
   _attn_qkvw->set_value((char*)(para_ptr + offset));
   _attn_qkvw->set_grad((char*)(grad_ptr + offset));
   offset += _hidden_size * _hidden_size * 3;
