+++ lightseq/csrc/layers_new/includes/transformer_encoder_layer.h
@@ -30,7 +30,7 @@ class TransformerEncoderLayer : public Layer {
 
   void before_backward() { return; }
 
-  int load_para_and_grad(const T1* para_ptr, T2* grad_ptr);
+  size_t load_para_and_grad(const T1* para_ptr, T2* grad_ptr);
 
   int load_params(const std::vector<const T1*>& para_vec, int offset);
 };
