+++ lightseq/csrc/layers_new/linear_layer.cpp
@@ -39,9 +39,9 @@ template <typename T1, typename T2>
 void LinearLayer<T1, T2>::before_backward() {}
 
 template <typename T1, typename T2>
-int LinearLayer<T1, T2>::load_para_and_grad(const T1* para_ptr,
-                                            T2* grad_ptr) {  // for training
-  int offset = 0;
+size_t LinearLayer<T1, T2>::load_para_and_grad(const T1* para_ptr,
+                                               T2* grad_ptr) {  // for training
+  size_t offset = 0;
 
   _linear_w->set_value((char*)(para_ptr + offset));
   _linear_w->set_grad((char*)(grad_ptr + offset));
