+++ lightseq/csrc/layers_new/multihead_attention_layer.cpp
@@ -113,9 +113,9 @@ void MultiheadAttentionLayer<T1, T2>::before_forward(size_t batch_size,
 }
 
 template <typename T1, typename T2>
-int MultiheadAttentionLayer<T1, T2>::load_para_and_grad(
+size_t MultiheadAttentionLayer<T1, T2>::load_para_and_grad(
     const T1* para_ptr, T2* grad_ptr) {  // for training
-  int offset = 0;
+  size_t offset = 0;
   _attn_qkvw->set_value((char*)(para_ptr + offset));
   _attn_qkvw->set_grad((char*)(grad_ptr + offset));
   _attn_qkvw->set_shape({3, _hidden_size, _hidden_size});
